Viz Thread Fix
GDK has its own internal mutex handling,

Rarely, what can happen is that the mx mutex is returned from draw_cb before the GDK mutex has been returned.

Basically when draw_cb is called it isnt called just by itself, it has a parent function that calls it. This parent function is the one that cannot be called at the same time gtk_widget_queue_draw_area is called.

So draw_cb is run the mutex is obtained, draw_cb finishes and the mutex is returned, but the parent function hasnt finised yet and we are still in an unsafe configuration even though mx is unlocked.

GTK has an internal drawing mutex that is triggered when the thread is safe this is the function (even though it is deprecated).